+++ src/Products/SiteAccess/__init__.py
@@ -1,0 +1,10 @@
+
+
+def initialize(context):
+    import VirtualHostMonster
+
+    context.registerClass(
+        instance_class=VirtualHostMonster.VirtualHostMonster,
+        permission='Add Virtual Host Monsters',
+        constructors=VirtualHostMonster.constructors,
+    )